adding users self editing and deleting

+++ src/main/java/org/peachSpring/app/controllers/UsersController.java
@@ -1,6 +1,7 @@
 package org.peachSpring.app.controllers;
 
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import org.peachSpring.app.exceptions.CannotDeleteUserException;
 import org.peachSpring.app.exceptions.UserNotFoundException;
@@ -11,7 +12,7 @@ import org.peachSpring.app.services.UserService;
 import org.peachSpring.app.util.constants.Gender;
 import org.peachSpring.app.util.search_config.UserSearchConfig;
 import org.peachSpring.app.util.search_config.constants.UserFilter;
-import org.peachSpring.app.util.validators.UserOfflineValidator;
+import org.peachSpring.app.util.validators.UserOnlineValidator;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -24,14 +25,15 @@ import java.util.Arrays;
 @RequestMapping("/users")
 public class UsersController {
 
-    private final UserOfflineValidator userOfflineValidator;
+    private final UserOnlineValidator userOnlineValidator;
     private final UserService userService;
     private final BooksUsersService booksUsersService;
     private final BookService bookService;
 
-    public UsersController(UserOfflineValidator userOfflineValidator, UserService userService, BooksUsersService booksUsersService, BookService bookService) {
+    public UsersController(UserOnlineValidator userOnlineValidator, UserService userService, BooksUsersService booksUsersService, BookService bookService) {
+        this.userOnlineValidator = userOnlineValidator;
 
-        this.userOfflineValidator = userOfflineValidator;
+
         this.userService = userService;
         this.booksUsersService = booksUsersService;
         this.bookService = bookService;
@@ -57,6 +59,7 @@ public class UsersController {
         model.addAttribute("stringToFind", searchConfig.getStringToFind());
         model.addAttribute("filter",searchConfig.getFilter());
         model.addAttribute("usersName", curUser.getName());
+        model.addAttribute("usersId", curUser.getId());
         return "users/index";
     }
     @GetMapping("/{id}")
@@ -68,9 +71,29 @@ public class UsersController {
             model.addAttribute("user", curUser);
             model.addAttribute("usersRole", principles.getRole());
             model.addAttribute("usersName", principles.getName());
+            model.addAttribute("usersId", principles.getId());
         } catch (UserNotFoundException e) {
             e.printStackTrace();
             return "errors/userNotFound";
+        }
+        return "users/user";
+    }
+    @GetMapping("/personal/{id}")
+    public String showPersonalData(@PathVariable("id") long id, Model model){
+        User curUser = userService.findOne(id);
+        User principles = UserService.getCurrentUsersPrinciples();
+        try {
+            model.addAttribute("id",id);
+            model.addAttribute("user", curUser);
+            model.addAttribute("usersRole", principles.getRole());
+            model.addAttribute("usersName", principles.getName());
+            model.addAttribute("usersId", principles.getId());
+        } catch (UserNotFoundException e) {
+            e.printStackTrace();
+            return "errors/userNotFound";
+        }
+        if (id!=principles.getId()){
+            return "redirect:/error";
         }
         return "users/user";
     }
@@ -88,7 +111,7 @@ public class UsersController {
                                 @ModelAttribute("newUser")
                                  @Valid User user,
                                 BindingResult bindingResult){
-        userOfflineValidator.validate(user,bindingResult);
+        userOnlineValidator.validate(user,bindingResult);
         if (bindingResult.hasErrors()){
             model.addAttribute("genders",Arrays.asList(Gender.values()));
             return "users/new";
@@ -99,10 +122,15 @@ public class UsersController {
     @GetMapping("/{id}/edit")
     public String requestToEditUser(Model model, @PathVariable("id") long id){
         User principles = UserService.getCurrentUsersPrinciples();
+        if (principles.getRole().equals("ROLE_USER")&& id!=principles.getId()){
+            System.out.println(id +" "+ principles.getId());
+            return "redirect:/error";
+        }
         model.addAttribute("usersRole", principles.getRole());
         model.addAttribute("usersName", principles.getName());
         model.addAttribute("genders",Arrays.asList(Gender.MALE,Gender.FEMALE));
         model.addAttribute("curUser", userService.findOne(id));
+        model.addAttribute("usersId", principles.getId());
         return "users/edit";
     }
     @PatchMapping("/{id}")
@@ -110,14 +138,30 @@ public class UsersController {
                             @ModelAttribute("curUser") @Valid User user,
                            BindingResult bindingResult,
                            @PathVariable("id") long id){
-        userOfflineValidator.validate(user,bindingResult);
-        model.addAttribute("genders",Arrays.asList(Gender.MALE,Gender.FEMALE));
+        User curUser = userService.findOne(id);
+        user.setPassword(curUser.getPassword());
+        user.setRole(curUser.getRole());
+        User principles = UserService.getCurrentUsersPrinciples();
+        userOnlineValidator.validate(user,bindingResult);
         if (bindingResult.hasErrors()){
+            model.addAttribute("genders",Arrays.asList(Gender.MALE,Gender.FEMALE));
+            model.addAttribute("usersRole", principles.getRole());
+            model.addAttribute("usersName", principles.getName());
+            model.addAttribute("usersId", principles.getId());
             return "users/edit";
         }
-        userService.update(id,user);
+        if (principles.getRole().equals("ROLE_USER")&& id!=principles.getId()){
+            System.out.println(id +" "+ principles.getId());
+            return "redirect:/error";
+        }
 
-        return "redirect:/users";
+        userService.update(id,user);
+        if (principles.getRole().equals("ROLE_USER")){
+            return "redirect:/users/personal/{id}";
+        } else {
+            return "redirect:/users/{id}";
+        }
+
     }
     @PatchMapping("/ban/{id}")
     public String banUser(Model model,
@@ -133,12 +177,25 @@ public class UsersController {
     }
 
     @DeleteMapping("/{id}")
-    public String deleteUser(@PathVariable("id") long id){
+    public String deleteUser(@PathVariable("id") long id,
+                             HttpServletRequest httpServletRequest){
+        User principles = UserService.getCurrentUsersPrinciples();
+        if (principles.getRole().equals("ROLE_USER")&& id!=principles.getId()){
+            System.out.println(id +" "+ principles.getId());
+            return "redirect:/error";
+        }
         try {
             userService.deleteById(id);
         } catch (CannotDeleteUserException e) {
             e.printStackTrace();
             return "errors/cannotDeleteUser";
+        }
+        if (principles.getRole().equals("ROLE_USER")){
+            HttpSession session = httpServletRequest.getSession(false);
+            if (session != null) {
+                session.invalidate();
+            }
+            return "redirect:/auth/login";
         }
         return "redirect:/users";
     }
